Remove redundant meta api
