--- cg/meta/microsalt/__init__.py
@@ -1,2 +1,0 @@
-""" export the lims microsalt api """
-from .lims import LimsMicrosaltAPI
